requirements.txt
fastapi[all]==0.111.1
uvicorn==0.30.3
mysql-connector-python==9.0.0
